+++ examples/serverless-framework/aws/package.json
@@ -47,6 +47,7 @@
     "prettier": "^1.5.2",
     "prettier-eslint": "^6.4.2",
     "serverless": "^1.17.0",
+    "serverless-apigwy-binary": "^0.1.0",
     "serverless-plugin-chrome": "^1.0.0-9"
   },
   "ava": {
