+++ examples/serverless-framework/aws/serverless.yml
@@ -10,6 +10,7 @@ provider:
 
 plugins:
   - serverless-plugin-chrome
+  - serverless-apigwy-binary  # For Raw Binary responses only
 
 custom:
   chromeFlags:
@@ -47,22 +48,44 @@ functions:
           path: version-info
           method: get
 
-  screenshot:
-    description: Headless Chrome Serverless-framework screenshot example
+  screenshot-html:
+    description: Headless Chrome Serverless-framework screenshot example (returns html)
     memorySize: 1536
     timeout: 30
-    handler: src/captureScreenshot.handler
+    handler: src/captureScreenshotHtml.handler
     events:
       - http:
           path: screenshot
           method: get
 
-  pdf:
-    description: Headless Chrome Serverless-framework pdf example
+  screenshot-raw:
+    description: Headless Chrome Serverless-framework screenshot example (returns binary)
     memorySize: 1536
     timeout: 30
-    handler: src/versionInfo.handler
+    handler: src/captureScreenshotBinary.handler
     events:
       - http:
-          path: version-info
+          path: screenshot
           method: get
+          contentHandling: CONVERT_TO_BINARY
+
+  pdf-html:
+    description: Headless Chrome Serverless-framework pdf example (returns html)
+    memorySize: 1536
+    timeout: 30
+    handler: src/printToPdfHtml.handler
+    events:
+      - http:
+          path: pdf
+          method: get
+
+  pdf-raw:
+    description: Headless Chrome Serverless-framework pdf example (returns binary)
+    memorySize: 1536
+    timeout: 30
+    handler: src/printToPdfBinary.handler
+    events:
+      - http:
+          path: pdf
+          method: get
+          contentHandling: CONVERT_TO_BINARY
